requirements.txt
django-bmemcached==0.2.4
django-heroku==0.3.1
Django==2.2.10
gunicorn==20.0.4
pylibmc==1.5.2
psycopg2-binary==2.8.4
dj_database_url==0.5.0
djangorestframework==3.10.3
